Rename property name & avoid Spark in tests
